LayoutTests/fast/events/file-input-hidden-in-ondrop-expected.txt

PASS dropTarget.value is "C:\\fakepath\\abe.png"
